requirements.txt
alembic==1.3.1
Babel==2.7.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dnspython==2.0.0
dominate==2.4.0
elasticsearch==7.6.0
email-validator==1.1.1
Flask==1.1.2
Flask-Admin==1.5.3
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Datepicker==0.13
Flask-Ext==0.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
numpy==1.19.0
pandas==1.0.5
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
six==1.13.0
SQLAlchemy==1.3.11
urllib3==1.25.8
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
